pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "jupyterlab",
    "jupyter",
    "diffusers",
    "ollama",
    "litellm",
    "sentence-transformers",
    "deepeval",
    "langgraph",
    "datasets",
    "bitsandbytes",
    "langdetect",
    "crawlee",
    "bertopic[visualization]>=0.17.3",
    "notebook>=7.4.5",
    "faiss-cpu>=1.12.0",
    "matplotlib>=3.10.5",
    "playwright>=1.55.0",
    "prefect>=3.4.17",
]
